Fixed homework
